requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
Django==4.0.5
djangorestframework==3.13.1
Flask==2.1.2
Flask-SQLAlchemy==2.5.1
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pytz==2022.1
requests==2.28.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
